consumer/requirements.txt
pika==1.2.1
psycopg2-binary==2.9.7
python-dotenv